useFilteredGroups hook
