+++ app/routes/$ws.forms._index/route.tsx
@@ -1,11 +1,10 @@
 import type { MetaFunction } from '@remix-run/node';
-import { useParams } from '@remix-run/react';
-import { useMemo, useState } from 'react';
 import { FormIdentifier, useCreateForm, useDeleteForm, useGroupedForms } from '~/data/form-api';
+import { HdBean } from '~/data/generated/openapi-dev';
 import { ProjectIdentifier } from '~/data/project-api';
 import { ArtifactCard, cardLinks, NewArtifactCard } from '~/neo/artifact/ArtifactCard';
 import { ArtifactGroup } from '~/neo/artifact/ArtifactGroup';
-import { filterArifacts, insertWorkspaceIfAbsent } from '~/neo/artifact/list-artifacts';
+import { useFilteredGroups } from '~/neo/artifact/useFilteredGroups';
 import { useNewArtifact } from '~/neo/artifact/useNewArtifact';
 import { useCreateEditor } from '~/neo/editors/useCreateEditor';
 import { Editor, useEditors } from '~/neo/editors/useEditors';
@@ -19,15 +18,8 @@ export const meta: MetaFunction = () => {
 };
 
 export default function Index() {
-  const { ws } = useParams();
   const { data, isPending } = useGroupedForms();
-  const [search, setSearch] = useState('');
-  const groups = useMemo(() => {
-    return insertWorkspaceIfAbsent(data ?? [], ws);
-  }, [data, ws]);
-  const filteredGroups = useMemo(() => {
-    return filterArifacts(groups, form => form.name.toLocaleLowerCase().includes(search.toLocaleLowerCase()));
-  }, [groups, search]);
+  const { filteredGroups, search, setSearch } = useFilteredGroups(data ?? [], (f: HdBean) => f.name);
   const { createFormEditor } = useCreateEditor();
   return (
     <Overview title='Forms' search={search} onSearchChange={setSearch} isPending={isPending}>
